feat: added reactive button coloring #4 moved login form to component file
