+++ frontend/app/assets/styles/bootstrap-custom.scss
@@ -1,6 +1,6 @@
 @import 'bootstrap/scss/functions';
 
-$primary: #06a0c6;
+$primary: #07b3df;
 $danger: #b82d2d;
 
 $font-family-base: 'Open Sans', sans-serif;
@@ -14,6 +14,9 @@ $font-family-base: 'Open Sans', sans-serif;
 :root {
   --bg-reactive-primary: #fcfcfc;
   --bg-reactive-secondary: #f5f5f5;
+  --bg-reactive-gray: #c7c7c7;
+  --bg-reactive-gray-hover: #{shade-color(#c7c7c7, 10%)};
+  --bg-reactive-gray-active: #{shade-color(#c7c7c7, 20%)};
   --text-reactive-primary: #111111;
   --text-reactive-secondary: #6e6e6e;
 }
@@ -21,6 +24,9 @@ $font-family-base: 'Open Sans', sans-serif;
 [data-bs-theme='dark'] {
   --bg-reactive-primary: #111111;
   --bg-reactive-secondary: #181818;
+  --bg-reactive-gray: #5e5e5e;
+  --bg-reactive-gray-hover: #{shade-color(#5e5e5e, 10%)};
+  --bg-reactive-gray-active: #{shade-color(#5e5e5e, 20%)};
   --text-reactive-primary: #fcfcfc;
   --text-reactive-secondary: #979797;
 }
@@ -28,12 +34,22 @@ $font-family-base: 'Open Sans', sans-serif;
 .bg-reactive-primary {
   background-color: var(--bg-reactive-primary);
 }
+
 .bg-reactive-secondary {
   background-color: var(--bg-reactive-secondary);
 }
+
+.btn-gray {
+  --bs-btn-bg: var(--bg-reactive-gray);
+  --bs-btn-hover-bg: var(--bg-reactive-gray-hover);
+  --bs-btn-active-bg: var(--bg-reactive-gray-active);
+  --bs-btn-color: var(--text-reactive-primary);
+}
+
 .text-reactive-primary {
   color: var(--text-reactive-primary);
 }
+
 .text-reactive-secondary {
   color: var(--text-reactive-secondary);
 }
